Added Binary File Support
