+++ Assets/EngyneCreations/Trello/Scripts/Trello.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "Trello"
+}
