+++ Assets/EngyneCreations/Trello/Scripts/Trello.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f0c8f74939bafdb41bcac436654e5d0b
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
