+++ Assets/EngyneCreations/Trello/Scripts/TrelloCard.cs
@@ -9,16 +9,44 @@
  * https://github.com/AdamEC/Unity-Trello
  */
 
-namespace Trello {
-    public class TrelloCard {
-	
+using UnityEngine;
+
+namespace Trello
+{
+    public class TrelloCard
+    {
+        public struct Attachment
+        {
+            private byte[] _fileSource;
+            private string _fileName;
+
+            public byte[] FileSource => _fileSource;
+            public string FileName => _fileName;
+
+            public Attachment(byte[] fileSource, string fileName)
+            {
+                _fileSource = fileSource;
+                _fileName = fileName;
+            }
+
+            public Attachment(Texture2D jpeg, string fileName)
+            {
+                _fileSource = jpeg.EncodeToJPG();
+                _fileName = fileName;
+            }
+
+            public bool IsValid()
+            {
+                return _fileSource != null && !string.IsNullOrEmpty(_fileName);
+            }
+        }
+        
         public string name = "";
         public string desc = "";
         public string due = "null";
         public string idList = "";
         public string urlSource = "null";
-        public byte[] fileSource = null;
-        public string fileName = null;
+        public Attachment attachment;
 
         /// <summary>
         /// Base class for a Trello card.
