+++ Assets/_BForBoss/_Core/Scripts/Core.asmdef
@@ -17,7 +17,8 @@
         "GUID:6ca18507fe0299a43bc059012bcfc822",
         "GUID:a26ed9ecc15b09a45b17fd3338273798",
         "GUID:145faacb266c53b4c9bddf4f754b5d09",
-        "GUID:0c752da273b17c547ae705acf0f2adf2"
+        "GUID:0c752da273b17c547ae705acf0f2adf2",
+        "GUID:f0c8f74939bafdb41bcac436654e5d0b"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
